% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasstrshiny-pkg.R
\docType{package}
\name{fasstrshiny-package}
\alias{fasstrshiny-package}
\alias{fasstrshiny}
\alias{fasstrshiny-pkg}
\title{fasstrshiny}
\description{
Package providing Shiny App user interface to the fasstr package.
}
\details{
Get started with \code{fasstr_shiny()}
}
